+++ back/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'timber-walle-e'
+rootProject.name = 'timber-wall-e'
