+++ back/src/main/java/com/coffefreaks/timberwalle/TimberWallEApplication.java
@@ -1,0 +1,12 @@
+package com.coffefreaks.timberwalle;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TimberWallEApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(TimberWallEApplication.class, args);
+	}
+}
