+++ back/src/main/java/com/coffefreaks/timberwalle/exception/TimberResourceNotFoundException.java
@@ -1,0 +1,20 @@
+package com.coffefreaks.timberwalle.exception;
+
+public class TimberResourceNotFoundException extends RuntimeException {
+
+    public TimberResourceNotFoundException() {
+        super();
+    }
+
+    public TimberResourceNotFoundException(final String message, final Throwable cause) {
+        super(message, cause);
+    }
+
+    public TimberResourceNotFoundException(final String message) {
+        super(message);
+    }
+
+    public TimberResourceNotFoundException(final Throwable cause) {
+        super(cause);
+    }
+}
