+++ back/src/main/java/com/coffefreaks/timberwalle/model/Circuit.java
@@ -1,0 +1,4 @@
+package com.coffefreaks.timberwalle.model;
+
+public class Circuit {
+}
