+++ back/src/main/java/com/coffefreaks/timberwalle/model/Enum/Direction.java
@@ -1,0 +1,8 @@
+package com.coffefreaks.timberwalle.model.Enum;
+
+public enum Direction {
+    DOWN,
+    UP,
+    LEFT,
+    RIGHT,
+}
