+++ back/src/main/java/com/coffefreaks/timberwalle/model/Request/CartographyRequest.java
@@ -1,0 +1,6 @@
+package com.coffefreaks.timberwalle.model.Request;
+
+import com.coffefreaks.timberwalle.model.Cartography;
+
+public class CartographyRequest extends Cartography {
+}
