+++ back/src/main/java/com/coffefreaks/timberwalle/model/Request/CircuitRequest.java
@@ -1,0 +1,6 @@
+package com.coffefreaks.timberwalle.model.Request;
+
+import com.coffefreaks.timberwalle.model.Circuit;
+
+public class CircuitRequest extends Circuit {
+}
