+++ back/src/main/java/com/coffefreaks/timberwalle/model/Request/LocationRequest.java
@@ -1,0 +1,6 @@
+package com.coffefreaks.timberwalle.model.Request;
+
+import com.coffefreaks.timberwalle.model.Location;
+
+public class LocationRequest extends Location {
+}
