+++ back/src/main/java/com/coffefreaks/timberwalle/model/Request/RobiotRequest.java
@@ -1,0 +1,31 @@
+package com.coffefreaks.timberwalle.model.Request;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class RobiotRequest {
+    private int id;
+    private String content;
+
+    public RobiotRequest(int id, String content) {
+        this.id = id;
+        this.content = content;
+    }
+
+    public String getContent() {
+        return content;
+    }
+
+    public void setContent(String content) {
+        this.content = content;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+}
