+++ back/src/main/java/com/coffefreaks/timberwalle/model/Response/CartographyResponse.java
@@ -1,0 +1,6 @@
+package com.coffefreaks.timberwalle.model.Response;
+
+import com.coffefreaks.timberwalle.model.Cartography;
+
+public class CartographyResponse extends Cartography {
+}
