+++ back/src/main/java/com/coffefreaks/timberwalle/model/Response/CircuitResponse.java
@@ -1,0 +1,6 @@
+package com.coffefreaks.timberwalle.model.Response;
+
+import com.coffefreaks.timberwalle.model.Circuit;
+
+public class CircuitResponse extends Circuit {
+}
