+++ back/src/main/java/com/coffefreaks/timberwalle/model/Response/LocationResponse.java
@@ -1,0 +1,11 @@
+package com.coffefreaks.timberwalle.model.Response;
+
+import com.coffefreaks.timberwalle.model.Location;
+
+public class LocationResponse extends Location {
+
+    public LocationResponse(Double positionX, Double positionY) {
+        this.setPositionX(positionX);
+        this.setPositionY(positionY);
+    }
+}
