+++ back/src/main/java/com/coffefreaks/timberwalle/model/Response/RobiotResponse.java
@@ -1,0 +1,46 @@
+package com.coffefreaks.timberwalle.model.Response;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class RobiotResponse {
+    private int id;
+    private String status;
+    private String content;
+
+    public RobiotResponse() {
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getStatus() {
+        return status;
+    }
+
+    public void setStatus(String status) {
+        this.status = status;
+    }
+
+    public String getContent() {
+        return content;
+    }
+
+    public void setContent(String content) {
+        this.content = content;
+    }
+
+    @Override
+    public String toString() {
+        return "RobiotResponse{" +
+                "id=" + id +
+                ", status='" + status + '\'' +
+                ", content='" + content + '\'' +
+                '}';
+    }
+}
