--- back/src/main/java/com/coffefreaks/timberwallee/TimberWalleEApplication.java
@@ -1,12 +1,0 @@
-package com.coffefreaks.timberwallee;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TimberWalleEApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TimberWalleEApplication.class, args);
-	}
-}
