--- back/src/main/java/com/coffefreaks/timberwallee/exception/TimberResourceNotFoundException.java
@@ -1,20 +1,0 @@
-package com.coffefreaks.timberwallee.exception;
-
-public class TimberResourceNotFoundException extends RuntimeException {
-
-    public TimberResourceNotFoundException() {
-        super();
-    }
-
-    public TimberResourceNotFoundException(final String message, final Throwable cause) {
-        super(message, cause);
-    }
-
-    public TimberResourceNotFoundException(final String message) {
-        super(message);
-    }
-
-    public TimberResourceNotFoundException(final Throwable cause) {
-        super(cause);
-    }
-}
