--- back/src/main/java/com/coffefreaks/timberwallee/model/Cartography.java
@@ -1,4 +1,0 @@
-package com.coffefreaks.timberwallee.model;
-
-public class Cartography {
-}
