--- back/src/main/java/com/coffefreaks/timberwallee/model/Enum/Direction.java
@@ -1,8 +1,0 @@
-package com.coffefreaks.timberwallee.model.Enum;
-
-public enum Direction {
-    DOWN,
-    UP,
-    LEFT,
-    RIGHT,
-}
