--- back/src/main/java/com/coffefreaks/timberwallee/model/Enum/EngineStatus.java
@@ -1,22 +1,0 @@
-package com.coffefreaks.timberwallee.model.Enum;
-
-public enum EngineStatus {
-    MOVING("moving"),
-    STOPPED("stopped"),
-    ERROR("internal error");
-
-    private final String label;
-
-    EngineStatus(String value) {
-        this.label = value;
-    }
-
-    public static EngineStatus valueOfLabel(String label){
-        for (EngineStatus e : values()) {
-            if (e.label.equals(label)) {
-                return e;
-            }
-        }
-        return null;
-    }
-}
