--- back/src/main/java/com/coffefreaks/timberwallee/model/Request/CartographyRequest.java
@@ -1,6 +1,0 @@
-package com.coffefreaks.timberwallee.model.Request;
-
-import com.coffefreaks.timberwallee.model.Cartography;
-
-public class CartographyRequest extends Cartography {
-}
