--- back/src/main/java/com/coffefreaks/timberwallee/model/Request/CircuitRequest.java
@@ -1,6 +1,0 @@
-package com.coffefreaks.timberwallee.model.Request;
-
-import com.coffefreaks.timberwallee.model.Circuit;
-
-public class CircuitRequest extends Circuit {
-}
