--- back/src/main/java/com/coffefreaks/timberwallee/model/Request/LocationRequest.java
@@ -1,6 +1,0 @@
-package com.coffefreaks.timberwallee.model.Request;
-
-import com.coffefreaks.timberwallee.model.Location;
-
-public class LocationRequest extends Location {
-}
