--- back/src/main/java/com/coffefreaks/timberwallee/model/Request/RobiotRequest.java
@@ -1,31 +1,0 @@
-package com.coffefreaks.timberwallee.model.Request;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class RobiotRequest {
-    private int id;
-    private String content;
-
-    public RobiotRequest(int id, String content) {
-        this.id = id;
-        this.content = content;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public void setContent(String content) {
-        this.content = content;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-}
