--- back/src/main/java/com/coffefreaks/timberwallee/model/Response/CartographyResponse.java
@@ -1,6 +1,0 @@
-package com.coffefreaks.timberwallee.model.Response;
-
-import com.coffefreaks.timberwallee.model.Cartography;
-
-public class CartographyResponse extends Cartography {
-}
