--- back/src/main/java/com/coffefreaks/timberwallee/model/Response/CircuitResponse.java
@@ -1,6 +1,0 @@
-package com.coffefreaks.timberwallee.model.Response;
-
-import com.coffefreaks.timberwallee.model.Circuit;
-
-public class CircuitResponse extends Circuit {
-}
