--- back/src/main/java/com/coffefreaks/timberwallee/service/Interface/CartographyService.java
@@ -1,19 +1,0 @@
-package com.coffefreaks.timberwallee.service.Interface;
-
-import com.coffefreaks.timberwallee.model.Request.CartographyRequest;
-import com.coffefreaks.timberwallee.model.Response.CartographyResponse;
-
-import java.util.List;
-
-public interface CartographyService {
-
-    List<CartographyResponse> getAll();
-
-    CartographyResponse create(CartographyRequest cartographyRequest);
-
-    CartographyResponse getById(Integer id);
-
-    CartographyResponse update(CartographyRequest cartographyRequest);
-
-    void delete(Integer id);
-}
