--- back/src/main/java/com/coffefreaks/timberwallee/util/RestPreconditions.java
@@ -1,41 +1,0 @@
-package com.coffefreaks.timberwallee.util;
-
-import com.coffefreaks.timberwallee.exception.TimberResourceNotFoundException;
-
-public class RestPreconditions {
-
-    private RestPreconditions() {
-        throw new AssertionError();
-    }
-
-    // API
-    /**
-     * Check if some value was found, otherwise throw exception.
-     *
-     * @param expression
-     *            has value true if found, otherwise false
-     * @throws TimberResourceNotFoundException
-     *             if expression is false, means value not found.
-     */
-    public static void checkFound(final boolean expression) {
-        if (!expression) {
-            throw new TimberResourceNotFoundException();
-        }
-    }
-
-    /**
-     * Check if some value was found, otherwise throw exception.
-     *
-     * @param resource
-     *            has value true if found, otherwise false
-     * @throws TimberResourceNotFoundException
-     *             if expression is false, means value not found.
-     */
-    public static <T> T checkFound(final T resource) {
-        if (resource == null) {
-            throw new TimberResourceNotFoundException();
-        }
-
-        return resource;
-    }
-}
