+++ back/src/main/resources/application.properties
@@ -14,7 +14,7 @@ robiot.endpoint.url=http://localhost:8080/ecotree/robiot-api/configurations
 
 # Logging
 logging.level.root=WARN
-logging.level.com.coffefreaks.timberwallee=TRACE
+logging.level.com.coffefreaks.timberwalle=TRACE
 
 application-description=@project.description@
 application-version=@project.version@
