+++ back/src/test/java/com/coffefreaks/timberwalle/TimberWalleEApplicationTests.java
@@ -1,0 +1,13 @@
+package com.coffefreaks.timberwalle;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TimberWalleEApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
