--- back/src/test/java/com/coffefreaks/timberwallee/TimberWalleEApplicationTests.java
@@ -1,13 +1,0 @@
-package com.coffefreaks.timberwallee;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TimberWalleEApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
