+++ back/src/test/resources/application.properties
@@ -3,4 +3,4 @@ robiot.endpoint.url=http://localhost:8080/ecotree/robiot-api/configurations
 
 # Logging
 logging.level.root=WARN
-logging.level.com.coffefreaks.timberwallee=TRACE
+logging.level.com.coffefreaks.timberwalle=TRACE
